+++ src/inmanta/compiler/help/templates/index_collision.j2
@@ -1,4 +1,4 @@
 The constructor `{{ constructor_str }}` ({{ constructor_loc }}) matches different instances in different indexes:
-{% for coll in collisions -%}
- - index {{ constructor_name }}({{ collisions[coll].index }}) matches {{ collisions[coll].constructor }}
+{% for index in collisions -%}
+ - index {{ constructor_name }}({{ index }}) matches {{ collisions[index] }}
 {% endfor %}
